feat(zetaclient): SUI deposits (#3534)
* Refactor and improve gateway event parsing

* fix module parsing

* Implement client.QueryModuleEvents

* typo fix

* Implement ObserveInbound

* Implement inbound trackers

* Add test cases for inbounds observation

* Add test cases for inbound trackers

* Update changelog

* fix event index

* Fix message constructor

* PR fixes

+++ zetaclient/chains/evm/observer/v2_inbound.go
@@ -201,7 +201,7 @@ func (ob *Observer) newDepositInboundVote(event *gatewayevm.GatewayEVMDeposited)
 		zetacore.PostVoteInboundCallOptionsGasLimit,
 		coinType,
 		event.Asset.Hex(),
-		event.Raw.Index,
+		uint64(event.Raw.Index),
 		types.ProtocolContractVersion_V2,
 		false, // currently not relevant since calls are not arbitrary
 		types.InboundStatus_SUCCESS,
@@ -341,7 +341,7 @@ func (ob *Observer) newCallInboundVote(event *gatewayevm.GatewayEVMCalled) types
 		zetacore.PostVoteInboundCallOptionsGasLimit,
 		coin.CoinType_NoAssetCall,
 		"",
-		event.Raw.Index,
+		uint64(event.Raw.Index),
 		types.ProtocolContractVersion_V2,
 		false, // currently not relevant since calls are not arbitrary
 		types.InboundStatus_SUCCESS,
@@ -485,7 +485,7 @@ func (ob *Observer) newDepositAndCallInboundVote(event *gatewayevm.GatewayEVMDep
 		1_500_000,
 		coinType,
 		event.Asset.Hex(),
-		event.Raw.Index,
+		uint64(event.Raw.Index),
 		types.ProtocolContractVersion_V2,
 		false, // currently not relevant since calls are not arbitrary
 		types.InboundStatus_SUCCESS,
